output: Own candidate buffer diff rendering
Operation candidate previews currently have a dedicated output diff module,
and candidate planning still exposes the raw buffer-diff renderer that
module calls.

The project does not provide a rendering boundary for the unified-diff text
itself. That leaves operation candidate planning coupled to difflib even
though the helper only formats preview output.

This commit addresses that by moving render_candidate_buffer_diff into
output.candidate_preview_diff and updating replacement preview rendering to
import the data helper.

A subsequent test commit will guard that the renderer stays out of batch
operation candidate planning.
