+++ src/git_stage_batch/batch/operation_candidates.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import difflib
 from dataclasses import dataclass
 from itertools import product
 from typing import Literal
@@ -373,26 +372,3 @@ def build_include_candidate_previews(
         )
         previews.append(preview)
     return tuple(previews)
-
-
-def render_candidate_buffer_diff(
-    file_path: str,
-    before_buffer: LineBuffer,
-    after_buffer: LineBuffer,
-    *,
-    label_before: str,
-    label_after: str,
-    context_lines: int,
-) -> str:
-    """Render a unified diff between two candidate buffers."""
-    before_text = before_buffer.to_bytes().decode("utf-8", errors="surrogateescape")
-    after_text = after_buffer.to_bytes().decode("utf-8", errors="surrogateescape")
-    return "".join(
-        difflib.unified_diff(
-            before_text.splitlines(keepends=True),
-            after_text.splitlines(keepends=True),
-            fromfile=f"{label_before}/{file_path}",
-            tofile=f"{label_after}/{file_path}",
-            n=context_lines,
-        )
-    )
